+++ CHANGELOG.md
@@ -1,20 +1,35 @@
 # Changelog
 
+## [1.1.0] - 2023-01-21
+
+**Features**
+
+- Read styles config from newProjectRoot #8
+
+**Testing**
+
+- Add workspace for new project root.
+- Open stories to page[ts|scss|html|spec].
+
 ## [1.0.2] - 2021-01-04
 
-**Features** 
+**Features**
+
 - Add read file from schematics root attribute or defaultProject #1
 
 **Testing**
+
 - Add workspace for default project.
 
 ## [1.0.1] - 2021-01-02
 
 **Fixes**
+
 - openStyleFile keybinding
 
 ## [1.0.0] - 2021-01-02
 
 **Features**
+
 - Switch files between typescript, style (css, scss), storybook, spec and html.
-- Detect style files from angular.json config. 
+- Detect style files from angular.json config.